Improved utility types
